🚧 WIP: Import Collection Resource
A user should be able to import a collection resource. In this commit, we are able to successfully import and create collection resources. From the console we can see the collection formed relatioships with works, but the frontend's count and display shows 0 relationships. Additionally, we are unable to re run the importer without receiving errors on the collection entry.

TODO: specs, refactor,

Issue:
- https://github.com/scientist-softserv/hykuup_knapsack/issues/161