♻️ Favor asking about model_name over class
Given our effort at lazy migration in Bulkrax we want to do a bit more
sniffing regarding the objects.  This is not quite adequate for the
general case of Collections but it is an improvement.

Ideally we should be interrogating the class and asking
`klass.collection?` but there are some confounding edge cases around
routing that we are in this pickle.

```ruby
irb(main):002:0> CollectionResource.model_name
=>
 @collection="collections",
 @element="collection",
 @human="Collection",
 @i18n_key=:collection,
 @klass=CollectionResource,
 @name="CollectionResource",
 @param_key="collection",
 @plural="collections",
 @route_key="collections",
 @singular="collection",
 @singular_route_key="collection">
irb(main):003:0> Collection.model_name
=>
 @collection="collections",
 @element="collection",
 @human="Collection",
 @i18n_key=:collection,
 @klass=Collection,
 @name="Collection",
 @param_key="collection",
 @plural="collections",
 @route_key="collections",
 @singular="collection",
 @singular_route_key="collection">
irb(main):004:0>
```