I161 import collection resources (#933)
* 🚧 WIP: Import Collection Resource

A user should be able to import a collection resource. In this commit, we are able to successfully import and create collection resources. From the console we can see the collection formed relatioships with works, but the frontend's count and display shows 0 relationships. Additionally, we are unable to re run the importer without receiving errors on the collection entry.

TODO: specs, refactor,

Issue:
- https://github.com/scientist-softserv/hykuup_knapsack/issues/161

* remove unused code

* refactor #conditionally_destroy_existing_files

This refactor was necessary because even though klass == ImageResource, which inherits from Valkyrie::Resouce through it's chain, klass === Valkyrie::Resource was returning false.

* exclude CollectionResource class from #destroy_existing_files

* WIP - try to import filesets with valkyrie resources

* Revert "WIP - try to import filesets with valkyrie resources"

This reverts commit 4ab31b69d6e0c584274978de74c9903e612836cd.

* 💄 rubocop fix

* i162 - import valkyrie works with filesets (#936)

* Revert "WIP - try to import filesets with valkyrie resources"

This reverts commit 4ab31b69d6e0c584274978de74c9903e612836cd.

* WIP

* WIP - try to import filesets with valkyrie resources

* 🚧 WIP: get filesets to import via bulkrax x valkyrie

* :tada: WIP: filesets to imports via bulkrax x valkyrie

There's still a lot to clean up here, but the import is successful in this commit.

* 💄 rubocop fixes

* uncomment #get_s3_files call and add collections to configuration

* Update object_factory.rb

* ♻️ Move method and remove single instance definition

I'm unclear why we were defining methods on the conf instance;
especially given that these exist on the configuration.

With this refactor, we're favoring using the Configuration object as the
container.

* Revert changes due to refactor coming in from main

* address errors post big refactor

* Refactoring for consistent method signatures

Also avoiding setting an unused instance variable

* :bug: remove passing user to work_resource add_file_sets and save merge to strategies

Importing a CSV of valkyrie works, collections, files and relationships is working at this point :tada:

* 🎁 Adding a new transaction step to handle different association

* ♻️ Extract update_index method to object factory

* ♻️ Extract object factory method

* ♻️ Extract add_resource_to_collection method

* ♻️ XIT out the mockery and stubbery of a spec

* ♻️ Extract method publish and add_child_to_parent_work

* ♻️ Rename method as it's not conditional

Yes, it is conditional but it operates on arrays that could be empty.

* Remove add to collection step

* 🐛 Fix publish parameter mismatch

* Removing custom transaction container.  We weren't using it

* Favor keyword args instead of hashes

* 💄fixing typo

* 🎁 Add update_collection to valkyrie object factory

* 💄 endless and ever appeasing of the coppers

---------

Co-authored-by: Jeremy Friesen <jeremy.n.friesen@gmail.com>

---------

Co-authored-by: Jeremy Friesen <jeremy.n.friesen@gmail.com>