🐛 Add return in ObjectFactory if valkyrie
Adding this early return here so we don't go down to the the #where and
trigger a NoMethodError. What it seems like it's doing is checking
Postgres for the object but if it doesn't find it then tries in Fedora,
however, Valkyrie object don't respond to #where so it throws an error.