Prevent metadata duplication (#969)
When the term Identifier is part of the metadata and used as source_identifier,
it resulted in the value getting put into the metadata twice. Calling uniq!
on the array removes any duplicate metadata, as there should never be a reason
for the same metadata to be on a record twice.