Revert ObjectFactory#find method returning false instead of nil on missing object (#1000)
the find method changed from returning nil to returning false in Hyrax 4 valkyrie support (#872). several bugs around this change have been guarded against, but a full audit was done and there is nothing that requires this method to continue to return false and several places that still expect it to be nil. So we are reverting to nil when an object isnt found in the factory