🐛 Fix `#importer_unzip_path` logic
We discovered a bug with the previous updates to
the `#importer_unzip_path` method where when the method is being called
to create a path, it returns `nil` because the path does not exist and
the whole reason why we are calling the method is to create the path.

We add a `mkdir` argument to differentiate between when we're looking up
the path vs wanting to create it.