Bulkrax should show errors when remote urls are not valid  (#1015)
* Report errors with files

Bulrax jobs were continuing without reporting an error, with nothing
appearing in the logs due to how the workers & logs interact. This
commit causes file errors to be reported in the importer even though
the job will end normally.

* ⚙️ Fix failing CI

Got the message `Gem::LoadError: pg is not part of
the bundle. Add it to your Gemfile.
(Gem::LoadError)` so this commit will add pg to the Gemfile.

---------

Co-authored-by: LaRita Robinson <larita@scientist.com>
Co-authored-by: Kirk Wang <kirk.wang@scientist.com>