Bulkrax editing without model (#1025)
* Support for CSV editing without a model

When editing via a CSV, if the model was omitted or incorrect, bulkrax
had been creating a new new object instead of updating the existing one.
This resulted in two Valkyrie resources with only the new one being
indexed.

This fix updates the `find_by_source_identifier` method (via
`find_by_property_value`) to now find without using the model, to ensure
that we find the existing record. It also updates the importer view to
display the actual model rather than finding so the displayed model
will be correct since may not always be present in the CSV.

* Appease rubocop