Fix error handling & event notifications
Valkyrie works were going through the save path in the valkyrie object
factory. This means the event notifications weren't being sent.

When this was fixed, it became clear the error handling in the
create_relationships_job was not correct, as the event jobs needed to
include a user. The relationship update was failing because it was
defined incorrectly.