Reload parent before saving
Prior work tried to carry along the updated parent
after a relationship was created, but as other jobs
could also be updating the parent, we could still lose
other changes while still keeping the relationship.

To be safer, we now re-fetch the parent immediately before
applying the relationship. This should ensure that we
always have the latest version of the parent.