Fix relationship race condition (#1028)
* Fix relationship race condition

https://github.com/samvera/bulkrax/issues/1027

Refactors the CreateRelationshipsJob to lock child records when necessary,
to ensure that concurrent jobs do not interfere with each other.

Fixes the save! method in the valkyrie_object_factory because the
resources were not being saved correctly since they respond to save...
they were being saved directly by calling save rather than by using
the persister.

* Fix error handling & event notifications

Valkyrie works were going through the save path in the valkyrie object
factory. This means the event notifications weren't being sent.

When this was fixed, it became clear the error handling in the
create_relationships_job was not correct, as the event jobs needed to
include a user. The relationship update was failing because it was
defined incorrectly.

* Turn on Create Relationships Job specs

Ever since valkyrization, the CreateRelationshipsJob specs have been
commented out. This commit rewrites them and turns them on.

* Rubocop appeasement

* Limit child works in relationships

Find based on identifier AND the correct importer run

* Fix race condition in CreateRelationshipsJob

The record locking does not seem to be working as expected. This commit
reloads the record immediately prior to updating it, to ensure that
the most recent version is used. This should help prevent lost data.

* Final spec fix for reworked relationship job

* Reload parent before saving

Prior work tried to carry along the updated parent
after a relationship was created, but as other jobs
could also be updating the parent, we could still lose
other changes while still keeping the relationship.

To be safer, we now re-fetch the parent immediately before
applying the relationship. This should ensure that we
always have the latest version of the parent.