Handle deletion & other issues with Valkyrie objects (#1049)
* Handle deletion & other issues with Valkyrie

- fixes method find_by_id This method was failing because it tried to use a class method on an instance. This caused a silent deletion failure, where the importer reported success but the object was not deleted. It also caused the show page to be broken.
- Refactors the FactoryClassFinder to add another way to find the class from an entry. The FactoryClassFinder did not find the class for deletions and other cases where the metadata was not parsed. This resulted in a fallback to the default work class.
- handles situations causing the entry show page to break

* Adjust specs for changes

Delete job now expects an error.
OAI entry spec needs to handle raw_metadata differently due to using it as a way to determine model.

* Rubocop indentation correction

* Do not mock raised error in deletion test

* Make delete job more flexible

Deletion required a source_identifier to exist, which was not always the case.
This change allows the delete job to work with entries that do not have a source_identifier,
by ensuring that the identifier is set based on the raw metadata if necessary.

* Adjust to get correct metadata for delete jobs

---------

Co-authored-by: Max Kadel <max@notch8.com>