Export files appropriately
Ensures that thumbnails and files are exported with filenames that
include the FileSet ID to avoid naming conflicts. This is especially
important when multiple files might have the same name.

Additionally, revises the logic related to mimetypes, as some calls to
Marcel::MimeType were did not include the necessary keyword arguments.

Moves some conditional logic into the object factories to keep the code cleaner.