Ensure work-fileset relationships
This change ensures that the relationships between works and their filesets
are recorded on both the work and the fileset during export. This ensures that
when importing into a new tenant, the relationships are preserved.

By adding a call to the parents method, the fileset row will have the
parent work's id.

However, this method is only available on Hyrax-based models that include the
ArResource module. If Bulkrax is used with other types of models, this logic
may need to be reworked into an object factory method that can handle different
types of models.