Exporter fixes (#1061)
* Fixes exporter bugs

1. Adds method for finding file sets from a valkyrized work to list of child methods.
2. Adds model_name method to the object factories and changes method for valkyrized works to find the correct models.
3. Fixes thumbnail handling... valkyrie uses thumbnail_id method.

* Fix logic to find thumbnails

* Fix use of internal_resource method

This is an instance method rather than a class method. Prior usage
resulted in incorrect class values.

* Expand specs to cover more exporter functionality

* Thumbnail method rename

* Validate all ValkyrieObjectFactory methods

* Export files appropriately

Ensures that thumbnails and files are exported with filenames that
include the FileSet ID to avoid naming conflicts. This is especially
important when multiple files might have the same name.

Additionally, revises the logic related to mimetypes, as some calls to
Marcel::MimeType were did not include the necessary keyword arguments.

Moves some conditional logic into the object factories to keep the code cleaner.

* Rubocop

* Ensure work-fileset relationships

This change ensures that the relationships between works and their filesets
are recorded on both the work and the fileset during export. This ensures that
when importing into a new tenant, the relationships are preserved.

By adding a call to the parents method, the fileset row will have the
parent work's id.

However, this method is only available on Hyrax-based models that include the
ArResource module. If Bulkrax is used with other types of models, this logic
may need to be reworked into an object factory method that can handle different
types of models.

* Address review feedback