Delete the file sets when we delete a work (#1064)
* Delete the file sets when we delete a work

We are not using transactions when deleting the
work, so the filesets were getting left behind.

* Fix exporting with missing files

In order to export & delete, we need to be able to handle works with missing filesets/files.

* Delete filesets when deleting works

Filesets must be deleted prior to the works they belong to.
Deletion jobs must have entry metadata built so the objects can be found.
Reimporting a deleted work should now works.

* Update app/models/concerns/bulkrax/export_behavior.rb

Co-authored-by: Max Kadel <max@notch8.com>

---------

Co-authored-by: Max Kadel <max@notch8.com>