Allow Hyrax to correctly handle sidebar menu
Bulkrax is attempting to support multiple versions of Hyrax, some of which
do not have the `sidebar_partials` method defined on `Hyrax::DashboardController`.
To ensure compatibility, we need to check for the existence of this method
before trying to use it.

Previously, the code was still overriding Hyrax's partial, resulting in
an incorrect sidebar menu display. By adding a conditional check, we can
ensure that Bulkrax only modifies the view paths when necessary, thus
allowing Hyrax to correctly handle the sidebar menu.